Fix compiler warnings from CI: enum-mixing sizer flags, dead code, GCC<12 attribute bug
compile_2204 (Ubuntu 22.04, GCC 11, plain -Wall -Wextra, no -Werror) was
the one CI job that actually surfaces warnings rather than failing outright
or building with -Werror. Fixed everything it reported:

- ~90x -Wdeprecated-enum-enum-conversion: sizer flag arguments OR'd
  wxAlignment/wxDirection/wxStretch constants directly (e.g.
  wxALIGN_CENTER_VERTICAL | wxALL), which C++20 deprecates. Cast the first
  flag in each OR-chain to int, which is enough to make every later `|` an
  int-enum operation instead of enum-enum, with no behavior change.
- ArtProvider.cpp's DecompressSvg() was reported unused: it's only called
  from the wxCHECK_VERSION(3, 1, 6) branch of GetImage(), so on the
  wxWidgets 3.0 this job builds against, the function itself compiles but
  its only caller doesn't. Wrapped the definition in the same #if guard.
- [[maybe_unused]] on a non-static data member triggers a spurious
  'attribute ignored' warning on GCC < 12 (reproduced directly against
  g++-11; fixed by g++-12), regardless of whether the member is actually
  used. Since Clang still needs the attribute for -Wunused-private-field,
  guarded it out only for old GCC via #pragma GCC diagnostic push/ignored,
  applied consistently to all three occurrences (Printout.h, SvgBitmap.h,
  wxMathml.h).

Verified with a full local build (all warnings gone, nothing new introduced)
under g++-11, a newer g++ with -Wall -Wextra -ansi -Werror -Wpedantic, and
clang++ -Wall -Wextra.

+++ src/dialogs/DiffFrame.cpp
@@ -476,7 +476,7 @@ toolBar->AddSeparator();
     });
 
     wxBoxSizer *paneSizer = new wxBoxSizer(wxVERTICAL);
-    paneSizer->Add(fileLabel, 0, wxEXPAND | wxALL, 2);
+    paneSizer->Add(fileLabel, 0, static_cast<int>(wxEXPAND) | wxALL, 2);
     wxBoxSizer *wsRow = new wxBoxSizer(wxHORIZONTAL);
     wsRow->Add(ws, 1, wxEXPAND);
     wsRow->Add(bar, 0, wxEXPAND);
