+++ src/dialogs/ResolutionChooser.cpp
@@ -49,14 +49,14 @@ ResolutionChooser::ResolutionChooser(wxWindow *parent, int id,
   wxFlexGridSizer *grid_sizer_1 = new wxFlexGridSizer(4, 2, 0, 0);
   wxBoxSizer *sizer_1 = new wxBoxSizer(wxHORIZONTAL);
   vsizer->Add(new wxStaticText(this, -1, _("The resolution for this image.")),
-              0, wxALL | wxALIGN_CENTER_HORIZONTAL, 5);
-  grid_sizer_1->Add(resolutionText, 0, wxALL | wxEXPAND, 5);
-  grid_sizer_1->Add(m_resolution, 0, wxEXPAND | wxLEFT | wxRIGHT, 2);
+              0, static_cast<int>(wxALL) | wxALIGN_CENTER_HORIZONTAL, 5);
+  grid_sizer_1->Add(resolutionText, 0, static_cast<int>(wxALL) | wxEXPAND, 5);
+  grid_sizer_1->Add(m_resolution, 0, static_cast<int>(wxEXPAND) | wxLEFT | wxRIGHT, 2);
 
   sizer_1->Add(button_1, 0, wxALL, 5);
   sizer_1->Add(button_2, 0, wxALL, 5);
   grid_sizer_1->Add(sizer_1, 1, wxALIGN_RIGHT, 0);
   SetAutoLayout(true);
-  vsizer->Add(grid_sizer_1, 1, wxEXPAND | wxLEFT, 0);
+  vsizer->Add(grid_sizer_1, 1, static_cast<int>(wxEXPAND) | wxLEFT, 0);
   SetSizerAndFit(vsizer);
 }
